refactor: 검색 결과 페이지 CardComponent 디자인 수정
