+++ src/pages/api/movies/search.ts
@@ -8,7 +8,7 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
       const query = req.query.query as string
       const page = (req.query.page as string) || '1' // 페이지 번호를 가져오거나 기본값 1을 사용합니다.
       const { data } = await axios.get(
-        `${API_URL}/search/movie?api_key=${API_KEY}&query=${query}&language=ko&page=${page}`,
+        `${API_URL}/search/multi?api_key=${API_KEY}&query=${query}&language=ko&page=${page}`,
       )
 
       res.status(200).json({ results: data.results })
